# node-api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Symptom Logger API",
    "version": "1.0.0",
    "description": "API documentation for the Symptom Logger API"
  },
  "servers": [
    {
      "url": "http://localhost:5200"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/symptom": {
      "post": {
        "summary": "Create a new symptom",
        "tags": [
          "Symptom"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Symptom created successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      },
      "get": {
        "summary": "Get all symptoms",
        "tags": [
          "Symptom"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of symptoms",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/symptom/{id}": {
      "put": {
        "summary": "Update an existing symptom",
        "tags": [
          "Symptom"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The symptom ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Symptom updated successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Symptom not found"
          }
        }
      },
      "get": {
        "summary": "Get a symptom by ID",
        "tags": [
          "Symptom"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The symptom ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Symptom details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Symptom not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a symptom by ID",
        "tags": [
          "Symptom"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The symptom ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Symptom deleted successfully"
          },
          "404": {
            "description": "Symptom not found"
          }
        }
      }
    },
...
  },
...
}